eigrpd: Make code a bit easier to read in eigrp_update.c
Lots of multiple pointer dereferences that made the
code lines really long, so let's properly setup
the pointers in a way that makes it easier to read.

No need to lookup the eigrp pointer because we already
have it, so let's use this( this will be good for the
future too when we have VRF's).

Signed-off-by: Donald Sharp <sharpd@cumulusnetworks.com>