zebra: Dissallow some static route options
Currently it is legal to specify a gateway and/or interface
with the `blackhole|null0|reject`.  As such:

`ip route 4.3.2.1/32 10.0.4.2 reject`
is a legal command to enter.

This has two issues:
1) Specifying either a gateway or an interface when attempting
to specify a blachole or reject route makes no sense
2) When entering the above command it is accepted, nvgenned
and silently ignored.

Modify both v4 and v6 code cli to not allow `reject|blackhole|null0`
to be accepted when you enter an nexthop or interface.  This
should solve the issues.

Signed-off-by: Donald Sharp <sharpd@cumulusnetworks.com>