zebra: Move selected fib route_entry to front of list
Move the selected fib route_entry to the front of the
rib_dest_t route_entry list( routes ).

The goal would be to allow for faster lookup of the
selected fib entry when we need to find it.  As
that it's a pretty common thing to do.

Additionally this has the side effect of modifying
'show ip route' behavior to show non FIB routes
before FIB routes.

Signed-off-by: Donald Sharp <sharpd@cumulusnetworks.com>

squash