lib: Allow free'd streams to be cached for next stream_new
streamcache -> Is an attempt to speed up buffer lookup
When we do a stream_free, if the size spot in streamcache
is not being used, then save the stream pointer
(clearing it up first).  Thus when we do a stream_new
lookup in the streamcache.

The belief here is that:
a) caching the size is faster then the FREE + MALLOC
   for the next one.
b) The daemon is going to allocate the same size
   more frequently than random sizes.

Recent perf tests show that stream_new and stream_free
were being called at a significant pace.

Signed-off-by: Donald Sharp <sharpd@cumulusnetworks.com>