bgpd: set timers correctly
When timers are manually configured differently from the default, we
never set the effective values on the peer. Instead we wait until the
negotiation step upon receipt of an OPEN message from the remote peer.

Among other things, this can cause us to wait the wrong amount of time
in OpenSent before closing the connection in case we never receive an
OPEN.

Signed-off-by: Quentin Young <qlyoung@cumulusnetworks.com>