tests: Add code to allow us to test packet handling in bgp easier
Fuzzing with tools like afl requires the ability to quickly/easily
call a function to test it's input.  In the case of bgp we
need the ability to test bgp's packet handler quickly.  This
added program bgpd/test_packet takes a file name as input
that it treats as normal bgp communication between itself
and a peer.

Signed-off-by: Donald Sharp <sharpd@cumulusnetworks.com>